Fixed make file for infinite scroll #2680
Ugh.  Two problems.  My views_infinite_scroll patches are conflicting with each other.  Moved to temporary.

Autopager was worse.  I got git to create a patch file that creates a whole new file for the css.  But drush make doesn't call git in a manner that creates files.  Since this is a library I can't just put in in temporary and I don't want to start another temporary folder for libs.  Instead I forked the lbirary.