#3265: Menu entries, basic workflow.
Created 2 menu entries. One is a local tab of /user, the other is
only hit when a new PIN user logs in to the system. The local tab
redirects to pinserver, and then either to a page to fill out the
rest of the user info, or their original destination.