Drush command downloads acquia domain settings.  #3090
Added os-domain-sync (os-da).  It gets acquia's domain settings and overwrites our domains.list file.  Useful after a file copy overwrites our dev settings with prod's list.

Added the command itself and help.  Also fixed the confirm file overwrite check to work with drush -y.  It used to loop, but now it will accept the changes.  This was necessary if we're going to run this from a post file copy hook.