Fixed fatal error in os_events #3530
`field_get_items` was being called with $field set to NULL.  That's where the fatal came from.  It was set to null because we were treating the date_id var like an array.  It was in fact a string, so instead of exploding the first element of the array, we were exploding the first letter of the string.  There was no way field could have been anything but null at this point.