First draft at db backup access as drupal page #3540
DB backups are listed on a drupal page now.  Hosting a page lets use skip copying all the backups since the server has access to where they're stored.  But it means we'll need to add a permission (and role?) for accessing them.

Acquia provides server vars that describe the host and env, so I'm able to find the backups no problem.

The only real question is where to put this.  I don't think it deserves its own module.  It's currently in OS, since that's a catch all for random pages.  But I don't think it's something that should be enabled for everyone.  Maybe if we had a more general version of iqss_scholar and iqss_projects that stored the things common to our installs, but not meant for anyone else?