Replace task completed callbacks with FALSE #3307
The completed property controls whether the task is run or not when a
group is inserted.

Since we always want our tasks to be run, the simplest way to do this
is to simply leave completed == FALSE. This is the default value, so we
don't even need to assign it.

Additionally, remove the now-unnecessary callback function definitions.