Access check, new module #3540
This didn't belong in os but wasn't worthy of a module of its own.  I made a harvard module for our customization that isn't part of os core.  I suspect some of the hwpi/iqss_scholar_front/iqss_projects_front modules could be moved in here.

Added an access check.  Didn't want to introduce a new role or permission, so it's just a var with a comma separated list of user names.