Fixed repec.  (no ticket)
==Problem==

Repec is old.  The repec service uses a perl script to scan a bunch of apache directories registered with the service.  That worked great on IQSS hosting, but on Acquia we don't have a spare directory sitting around, so there's nothing to scan.

Originally I wanted to save the files to a magic directory and give a shell script passwordless access to rsync that directory.  Yeah.

==Solution==

Instead, I made a path that replicates the repec directory.  It lists its series and archive rdfs as well as all the repec enabled papers.  Clicking on any of those generates the file via the same code that used to save files.  I'm tempted to point repec to this directory, but I'm not really sure what their script does so instead I'm going to use wget -m to mirror the new path and save that to the original directory we used in 6.  This can happen in a cron job and I've supplied a shell script.