Only get fn/ln from users.  Concatenate domains.  #3925
Didn't realize the name fields were attached to profiles as well.  uid 178 was also getting names from node 178 when that node was a person.

Domains are space separated and added to end of line.  This probably could have been part of the original query but I didn't see the point in a one off script.