Vsite domain works with select form #3887
When the form used a text field $form_state['values'] included the value entered there.  When I switched to a select, the value was absent from values, but it was input.  I tried just using that in the submit function.  That wasn't quite enough.  Turns out there was another submit function after mine, provided by spaces.  It was saving the values and kept getting a blank because of the values/inputs problem.  Copying from input into value worked.  I'm still not sure why this was necessary...