Loading gif for getsatsifaction support #3961
Getsatisfaction appends its iframe to a div rather than replacing the div.  I couldn't just leave throbber.gif in there and let it disappear once the iframe loaded.  Instead I used css.  All elements except the last element in the div are hidden.  When the throbber is alone its visible.  When it has an iframe after it, only the iframe is visible.