cp/menu add handles external vsites better #3663
The cp menu wizard tries to offer the ability to create a new node when a url is entered that does not exist.  It failed to take into account other vsites.  Now it makes sure the url is valid for the active vsite before offering a new node.  If the url points to an external vsite, just use the url.