cp menu validation state bug #3663
While testing #3663 I noticed that when a menu item failed validation, it sometimes kept the values from before the first pass at validating.  ie, create a menu entry with no title.  It fails because it needs a title.  Add a title and the form passes, but is saved with no title.

Form values were being cached between pages, but they weren't being updated with the new values before being saved for the next page.  This was fixed by swapping an array addition for an array_merge.