Fixed javascript for display styles #2033
This was so much easier to manage with a definitive list of view modes per bundle.  I sent that setting to Drupal.settings and let javascript work out what was available each time the content type changed.  So easy.

sort and more links do similar things, but I think it'd be nice to make them a bit more consistent.