Started fixing sv list repeat events #2033
I cut down the three functions from the old box into one function, that's smaller than the smallest of the original three.  Basically I'm taking the array of event nodes and exploding it on the date field.  So if a single node has 5 dates, it'll be replaced with 5 instances of that node, each with different dates.

... but `entity_view` calls `entity_key_array_by_property` which reduces that group of arrays I made into a list of unique nids.  I think this is the first case I've found for a custom view function instead of entity_view.  Oh well.