SV list supports repeating events #2033
I was mostly able to keep this in the event plugin.  I added a render hook to the plugins.  It gives each plugin a chance to render entities.  Only if it returns no results does the default render run.

Using a custom render function, the events plugin skips past the entity_load and just does a node_view on each of the event nodes.  That's what we did in the old box.

I was able to keep the loop that split nodes with repeat rules into several copies of the same node (see prev commit).  I had to add a date_id key for each delta.  I also keyed the nodes by event timestamp, so they could sort easily after being split up.