Re-ordered sorting, specific options go to top #2033
When a content type specific sort is present, it's not because it's a weird excpetion.  It's because the default sorts didn't make sense.  These sorts should be more visible.  I moved them to the top of the list and the generics are at the bottom.

I still want to see if I can make newest/oldest/alpha/random something that the entity sv list can sort instead of relying on node and file to implement them