Made title a view mode.  #2033
Title is a view mode.  So is links, for link nodes.

Confirmed title sv lists work across all content types.  By default it worked, but the view modes that had customization were problematic.  All the items modified in this commit needed minor changes.  So does feed_importer, but I'm less worried about that since we're probably removing it from the SV lis anyway.

Also reimplemented teh links view mode for links.  So if you select title, you'll get the title of the ndoe linking ot the node.  If you select links as your display mode, you'll get the title of the node linking to the desintation.