Added term select to list of files.  #4502
This wasn't too bad so far.  I had to update the helper function that got attached vocabs to use an entity type instead of being hardcoded to node.  I also had to tell each plugin what the entity type was so the helper funciton could be called correctly.  Instead of defining this in each of their __construct's, I had them extend a parent class that set the var.  Then I went back and made each file extend it.  Not really a shortcut after all, but I think this will be useful long term and it gives me a single point of entry for documenting all the various hook methods I've created.