LOF filters by term #4502
Changed all the node specific options in the tagged node filter target entities instead.

The only real change is that the taxonomy_index table only handles tagged nodes.  I switched to og's field_data_og_vocabulary instead to get to tagged files.  There's no functional difference, but this table is a bit bigger.

Still need to test how it works with multiple vocabs.  I simplified the code a little in this case so I'm not 100% sure it's doing the same thing yet.