Preload sv taxonomy/bundle data #2033
The ajax call to fetch vocabs containing a bundle was buggy.  When aggregation was enabled, it was blocking the other javascript from running, which is why content specific options were greyed out.

Instead of fixing it, I preloaded the list of vids/bundles into Drupal.settings.  I'd been meaning to do this for a while so you don't need to wait for ajax when the content type changes.  Sort and display styles already work this way, so I didn't see why taxonomy should be the exception.